feat: Implement add-in function
The project type for creating add-in should be Class Library, instead of Windows Forms App. Therefore, a new Class Library project was created.
The original project (NoteQuickFormatter) was kept to ensure this commit remains clean and focused on the new add-in implementation. It will be removed in a future commit.